# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OM-Paie API",
        "description": "API de paiement mobile et transfert d'argent - OM-Paie avec MongoDB",
        "contact": {
            "email": "support@om-paie.sn"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://tinkin-transfer.onrender.com/api",
            "description": "Serveur de production sur Render"
        },
        {
            "url": "http://localhost:8000/api",
            "description": "Serveur de développement"
        }
    ],
    "paths": {
...
        "/transfer": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Effectuer un transfert d'argent",
                "description": "Effectue un transfert d'argent entre l'utilisateur authentifié et un destinataire",
                "operationId": "transfer",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "telephone",
                                    "montant"
                                ],
                                "properties": {
                                    "telephone": {
                                        "description": "Numéro de téléphone du destinataire",
                                        "type": "string",
                                        "example": "+221781234567"
                                    },
                                    "montant": {
                                        "description": "Montant à transférer (min: 1, max: 1000000)",
                                        "type": "number",
                                        "format": "float",
                                        "example": 5000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Transfert effectué avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Transfert effectué avec succès"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "64f7b1a2c5d4e123456789ad"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "transfer"
                                                },
                                                "montant": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 5000
                                                },
                                                "expediteur_telephone": {
                                                    "type": "string",
                                                    "example": "+221771234567"
                                                },
                                                "destinataire_telephone": {
                                                    "type": "string",
                                                    "example": "+221781234567"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "completed"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur métier (solde insuffisant, etc.)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Non authentifié",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponse": {
                "title": "Réponse d'erreur",
                "description": "Réponse standard en cas d'erreur",
                "properties": {
                    "success": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "type": "string",
                        "example": "Message d'erreur"
                    },
                    "error": {
                        "type": "string",
                        "example": "error_code",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "title": "Erreur de validation",
                "description": "Réponse en cas d'erreur de validation",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "properties": {
                            "field_name": {
                                "type": "array",
                                "items": {
                                    "type": "string",
                                    "example": "Le champ field_name est requis."
                                }
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "OAuth2 avec Laravel Passport",
                "flows": {
                    "password": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            },
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token (JWT) obtenu après connexion",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}